Move backend and frontend files
