+++ user-pages/chauncey-den/chauncey-den.html
@@ -5,7 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <link href="../../style.css" rel="stylesheet">
     <link href="./chauncey-den.css" rel="stylesheet">
-    <script type="module" src="./chauncey-den.js"></script>
+    <script type="module" src="./frontend/chauncey-den.js"></script>
     <title>Chauncey's Den</title>
 </head>
 <body id="page-container">
